Laget egen mappe for visningsvarianter og renamed metadata
