+++ src/main/kotlin/no/nav/familie/pdf/pdf/PdfUtils.kt
@@ -34,10 +34,10 @@ import no.nav.familie.pdf.pdf.PdfElementUtils.lagOverskriftH2
 import no.nav.familie.pdf.pdf.PdfElementUtils.lagOverskriftH3
 import no.nav.familie.pdf.pdf.PdfElementUtils.lagVerdiElement
 import no.nav.familie.pdf.pdf.PdfElementUtils.navLogoBilde
-import no.nav.familie.pdf.pdf.VisningsvariantUtils.håndterVisningsvariant
 import no.nav.familie.pdf.pdf.domain.FeltMap
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
 import no.nav.familie.pdf.pdf.språkKonfigurasjon.SpråkKontekst
+import no.nav.familie.pdf.pdf.visningsvarianter.håndterVisningsvariant
 import org.slf4j.LoggerFactory
 
 object PdfUtils {
@@ -71,7 +71,7 @@ object PdfUtils {
         val sideantallInnholdsfortegnelse =
             if (harInnholdsfortegnelse) kalkulerSideantallInnholdsfortegnelse(feltMap, innholdsfortegnelse) else 0
 
-        UtilsMetaData.leggtilMetaData(pdfADokument, feltMap)
+        leggtilMetaData(pdfADokument, feltMap)
 
         Document(pdfADokument).apply {
             settFont(FontStil.REGULAR)
